requirements.txt
streamlit
azure-ai-formrecognizer
azure-core
openai
python-dotenv
reportlab
numpy